Added Manual Database Additions
This commit adds the "linkUser" command

This command lets an administrator manually link a person's discord and uw ID.
This is particularly useful if the database is wiped as now you can add people back.
The command even checks to see if the person is already verified and acts accordingly!